+++ Prototype2/src/main/java/com/group7/prototype/repository/ItemRepository.java
@@ -12,7 +12,7 @@ public interface ItemRepository {
 
     Item findItemByBarcode(int barcode);
 
-    Item editItem(Integer barcode, String name, Integer quantity, Integer lQuantity, Float price);
+    Item editItem(Integer barcode, String name, Integer quantity, Integer lQuantity, Double price);
 
     boolean deleteItem(int barcode);
 }
